ready to launch

+++ Domain/Interfaces/Clients/IProductStockRepository.cs
@@ -5,7 +5,7 @@ namespace Domain.Interfaces.Clients
 {
     public interface IProductStockRepository : IGenericRepository<ProductStocks, NameBaseEntityFilter>
     {
-        Task UpdateMinStock(int productId, int quantity, string dbName);
-        Task UpdateAddStock(int productId, int quantity, string dbName);
+        Task UpdateMinStock(int productId, double quantity, string dbName);
+        Task UpdateAddStock(int productId, double quantity, string dbName);
     }
 }
